fix(impl-review-r1-T1-1): revert in-place migration edits and i64 wire widen [CRITICAL #7, HIGH #1]
Addresses 06-implementation-review.md Round 1 CRITICAL #7 and HIGH #1.
Pre-fix state: pre-impl-review/round-1/pr-13.

CRITICAL #7 — restore the original `DEFAULT 'ISS'` in the two pre-existing
applied migrations so engineer DBs already on those checksums are not
broken by an in-place edit:

  - crates/db/migrations/20260502120000_create_organizations.sql
  - crates/remote/migrations/20260112000000_remote-projects.sql

The synthetic-identity seeder (PR #16) writes `issue_prefix = 'VK'`
explicitly, and `derive_issue_prefix` returns 'VK' for the default-fallback
path, so the column DEFAULT is functionally moot — no replacement
forward-only migration is needed.

HIGH #1 — revert the i32→i64 widen of `Issue.issue_number` and the new
`widen-issue-number-to-bigint.sql` Postgres migration. The widen is
out-of-scope for Subtask 1.3 (the boundary is the local `crates/db`
SQLite store) and contradicts Subtask 2.1 "no behavioral change" on the
wire. Local i64 storage is preserved; the wire/api-types stays i32 with
the `as i32` narrowing at the model→wire boundary.

+++ crates/db/src/models/issue.rs
@@ -365,7 +365,7 @@ impl From<Issue> for wire::Issue {
         Self {
             id: value.id,
             project_id: value.project_id,
-            issue_number: value.issue_number,
+            issue_number: value.issue_number as i32,
             simple_id: value.simple_id,
             status_id: value.status_id,
             title: value.title,
